TickeType rest vastauket ja dokumentaatio

+++ ticketguru/src/main/java/spagetti/tiimi/ticketguru/web/TicketTypeRestController.java
@@ -3,15 +3,19 @@ package spagetti.tiimi.ticketguru.web;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
+import spagetti.tiimi.ticketguru.Exception.BadRequestException;
+import spagetti.tiimi.ticketguru.Exception.NotFoundException;
 import spagetti.tiimi.ticketguru.domain.TicketType;
 import spagetti.tiimi.ticketguru.domain.TicketTypeRepository;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.PutMapping;
 
 @RestController
@@ -19,37 +23,59 @@ public class TicketTypeRestController {
 
     private TicketTypeRepository trepository;
 
-    public  TicketTypeRestController (TicketTypeRepository trepository) {
+    public TicketTypeRestController(TicketTypeRepository trepository) {
         this.trepository = trepository;
     }
 
     @GetMapping("/tickettypes")
+    @ResponseStatus(HttpStatus.OK)
     public List<TicketType> ticketTypesRest() {
         return (List<TicketType>) trepository.findAll();
     }
 
     @GetMapping("/tickettypes/{id}")
+    @ResponseStatus(HttpStatus.OK)
     public Optional<TicketType> getTicketTypeRest(@PathVariable Long id) {
+        if (!trepository.findById(id).isPresent()) {
+            throw new NotFoundException("TicketType does not exist");
+        }
         return trepository.findById(id);
     }
 
     @PostMapping("/tickettypes")
+    @ResponseStatus(HttpStatus.CREATED)
     public TicketType newTicketTypeRest(@RequestBody TicketType ticketType) {
+        if (ticketType.getTypeid() != null) {
+            throw new BadRequestException("Do not include typeid");
+        }
+        if (ticketType.getName().isEmpty()) {
+            throw new BadRequestException("TicketType name is missing");
+        }
         return trepository.save(ticketType);
     }
-    
+
     @DeleteMapping("tickettypes/{id}")
+    @ResponseStatus(HttpStatus.OK)
     public void deleteTicketTypeRest(@PathVariable Long id) {
+        if (!trepository.findById(id).isPresent()) {
+            throw new NotFoundException("TicketType does not exist");
+        }
         trepository.deleteById(id);
     }
 
     @PutMapping("tickettypes/{id}")
+    @ResponseStatus(HttpStatus.CREATED)
     public Optional<TicketType> editTicketTypeRest(@PathVariable Long id, @RequestBody TicketType updatedTicketType) {
+        if (!trepository.findById(id).isPresent()) {
+            throw new NotFoundException("TicketType does not exist");
+        } else if (updatedTicketType.getName().isEmpty()) {
+            throw new BadRequestException("TicketType name is missing");
+        }
         return trepository.findById(id)
-            .map(ticketType -> {
-                ticketType.setName(updatedTicketType.getName());
-                ticketType.setNote(updatedTicketType.getNote());
-                return trepository.save(ticketType);
-            });
+                .map(ticketType -> {
+                    ticketType.setName(updatedTicketType.getName());
+                    ticketType.setNote(updatedTicketType.getNote());
+                    return trepository.save(ticketType);
+                });
     }
 }
